+++ src/core/bg-physics.js
