+++ src/core/boot-overlay.js
